style giude fixed, and some minor corrections
