+++ src/main/java/com/codeborne/selenide/hookperformers/FlashAction.java
@@ -1,0 +1,45 @@
+package com.codeborne.selenide.hookperformers;
+
+import org.openqa.selenium.WebElement;
+
+import static com.codeborne.selenide.Configuration.*;
+import static com.codeborne.selenide.Selenide.executeJavaScript;
+import static com.codeborne.selenide.Selenide.sleep;
+
+public class FlashAction extends MarkAction {
+
+  private int flashPause = 200;
+
+  @Override
+  public boolean conditionForAction(WebElement element, String methodName) {
+    return presentationMode.active && presentationMethods.contains(methodName) && presentationMode.flashElements;
+  }
+
+  @Override
+  public void action(WebElement element, String methodName) {
+
+    sleep(presentationMode.delayBeforeCommand);
+
+    String flasherId = "selenideFlasher";
+    switch (methodName) {
+      case "doubleClick":
+        for (int i = 0; i < 2; i++) {
+          markElement(element, presentationMode.flashColor, flasherId);
+          sleep(flashPause);
+          removeMarker(flasherId);
+          sleep(flashPause);
+        }
+        break;
+      default:
+        markElement(element, presentationMode.flashColor, flasherId);
+        sleep(flashPause);
+        removeMarker(flasherId);
+        break;
+    }
+  }
+
+  private static void removeMarker(String elementId) {
+    executeJavaScript("if (document.contains(document.getElementById('" + elementId + "'))) {" +
+            "document.getElementById('" + elementId + "').remove()}");
+  }
+}
