+++ src/main/java/com/codeborne/selenide/hookperformers/HookAction.java
@@ -1,0 +1,15 @@
+package com.codeborne.selenide.hookperformers;
+
+import org.openqa.selenium.WebElement;
+
+public interface HookAction {
+  /**
+   * Condition for activation
+   */
+  boolean conditionForAction(WebElement element, String methodName);
+
+  /**
+   * Action when activated
+   */
+  void action(WebElement element, String methodName);
+}
