+++ src/main/java/com/codeborne/selenide/impl/SelenideElementProxy.java
@@ -4,7 +4,7 @@ import com.codeborne.selenide.SelenideElement;
 import com.codeborne.selenide.commands.Commands;
 import com.codeborne.selenide.ex.InvalidStateException;
 import com.codeborne.selenide.ex.UIAssertionError;
-import com.codeborne.selenide.hookPreformers.HookPerformer;
+import com.codeborne.selenide.hookperformers.HookPerformer;
 import com.codeborne.selenide.logevents.SelenideLog;
 import com.codeborne.selenide.logevents.SelenideLogger;
 import org.openqa.selenium.InvalidElementStateException;
